Move the user api methods into a separate controller in the api namespace
